Fix precedence, argument width and staging in the I/O syscalls
Re-audit of the previous two commits, asking of each change not "is the
errno right" but "what does the accepted value then do, and in what order".
Six answers were wrong.

Precedence. A check is only correct at the point the kernel runs it:

  preadv2/pwritev2  `vfs_readv` reaches `kiocb_set_rw_flags` only after the
                    FMODE check, after `import_iovec`, and after the
                    `if (!tot_len) goto out;` short-circuit. Validating the
                    flag word first reported -EOPNOTSUPP where Linux reports
                    -EBADF for a closed fd, or 0 for an empty vector.
  fcntl F_DUPFD     `do_fcntl` is handed an already-resolved `struct file *`,
                    so a closed descriptor is -EBADF before the floor is ever
                    inspected. The new floor check pre-empted it.
  F_ADD_SEALS       `memfd_add_seals` validates the requested seal set before
                    it looks at what the file already carries, so an
                    undefined bit is -EINVAL even on an already-sealed memfd.
                    An `allow_sealing` early-return answered -EPERM instead —
                    and was redundant besides, since a memfd created without
                    MFD_ALLOW_SEALING already starts with F_SEAL_SEAL.

Argument width. The register is not the argument:

  eventfd/eventfd2  `SYSCALL_DEFINE2(eventfd2, unsigned int, count, ...)` —
                    the initial counter is 32 bits. Seeding from the full
                    register starts `eventfd2(1 << 32 | 5, 0)` at 4294967301
                    instead of 5, so the first read returns a value that
                    never existed.
  fcntl F_DUPFD     `int argi = (int)arg` — the floor is the low 32 bits.
                    `fcntl(fd, F_DUPFD, 1 << 32)` is a floor of 0 and
                    duplicates; comparing the untruncated register against
                    RLIMIT_NOFILE rejected it with -EINVAL.

Staging. `validate_rw_user_range` deliberately skips NARF's generic 16-MiB
single-copy cap so a large request is not rejected outright, and that only
works if the handler then honours the cap by chunking — which is why
sys_read and sys_write stage at 64 KiB. The positional rewrites did not
carry that over: pread64 allocated a buffer the size of the whole request,
up to MAX_RW_COUNT, then failed the copy on the 16-MiB limit; pwrite64 and
the preadv/pwritev core had the same shape, and a SINGLE oversized iovec
slipped through the vector-wide MAX_RW_COUNT cap. All four now stage in
bounded chunks, so a transfer past the limit completes as it does on Linux
instead of asking the kernel heap for 2 GiB and then reporting -EINVAL.

Also: `do_epoll_create` opens its anon inode `O_RDWR | (flags & O_CLOEXEC)`.
NARF left the access mode at O_RDONLY, which glibc's stdio observes through
F_GETFL when an fd is handed to fdopen.

Ten more tests, covering each precedence pair, both truncations, and the
three staging paths. The chunked-write test reads back the far end of the
transfer, so a short write that reported the full count cannot pass.

Targeted syscall_abi + userspace: x86_64 1819 pass, aarch64 1673 pass.

+++ userspace/src/handlers/sys_eventfd.rs
@@ -52,9 +52,15 @@ fn create_eventfd(ctx: &mut dyn TrapContext, initval: u64, flags: u32) {
 
 /// `eventfd2(initval, flags)` — x86_64 290, aarch64 19. The form every libc
 /// `eventfd()` wrapper actually issues.
+///
+/// `SYSCALL_DEFINE2(eventfd2, unsigned int, count, int, flags)`: the initial
+/// counter is a 32-BIT argument. Reading the register as a full u64 seeded a
+/// counter Linux would have truncated — `eventfd2(1 << 32, 0)` starts at zero
+/// there and at 4294967296 here, so the first read returns a value that never
+/// existed and the semaphore drains 2^32 times instead of not at all.
 pub(crate) fn sys_eventfd2(ctx: &mut dyn TrapContext) {
     let args = *ctx.args();
-    create_eventfd(ctx, args.arg0, args.arg1 as u32);
+    create_eventfd(ctx, u64::from(args.arg0 as u32), args.arg1 as u32);
 }
 
 /// `eventfd(initval)` — the original one-argument call, x86_64 284 only.
@@ -66,6 +72,7 @@ pub(crate) fn sys_eventfd2(ctx: &mut dyn TrapContext) {
 /// CLOEXEC or NONBLOCK descriptor it never asked for, or -EINVAL, entirely
 /// according to register garbage.
 pub(crate) fn sys_eventfd(ctx: &mut dyn TrapContext) {
-    let initval = ctx.args().arg0;
+    // `SYSCALL_DEFINE1(eventfd, unsigned int, count)` — 32-bit, as above.
+    let initval = u64::from(ctx.args().arg0 as u32);
     create_eventfd(ctx, initval, 0);
 }
